bugfix epkg init error

+++ lib/epkg/repo.sh
@@ -9,7 +9,7 @@ init_channel_repo()
 	local repo=$3
 
 	# channel.yaml
-	[[ -f $EPKG_CACHE/epkg-manager/channel/${channel}-channel.yaml ]] || echo "channel ${channel} not found" && return 1
+	[[ -f $EPKG_CACHE/epkg-manager/channel/${channel}-channel.yaml ]] || { echo "channel ${channel} not found" && return 1; }
 	local env_channel_yaml=${HOME}/.epkg/envs/${env}/profile-current/etc/epkg/channel.yaml
 	mkdir -p $(dirname ${env_channel_yaml})
 	cp $EPKG_CACHE/epkg-manager/channel/${channel}-channel.yaml  $env_channel_yaml
